+++ app/Http/Controllers/CourseController.php
@@ -15,6 +15,9 @@ class CourseController extends Controller
         return view('cms.courses.index')->with('courses',$courses);
     }
     public function add(Request $request){
+        if(!in_array('Add Course',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $mentors = User::getUserByRole('Teacher');
         $data = [
             'mentors' => !empty($mentors) ? $mentors : [], 
@@ -85,6 +88,9 @@ class CourseController extends Controller
     }
 
     public function courseEdit(Request $request, $id) {
+        if(!in_array('Edit Course',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $course = Course::find($id);
         $mentors = User::getUserByRole('Teacher');
         $data = [
@@ -138,6 +144,9 @@ class CourseController extends Controller
 
     public function destroy($id)
     {
+        if(!in_array('Delete Course',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $course = Course::find($id);
         $course->delete();
         return redirect()->back()->with('msg', 'Course Deleted Successfully!');
