+++ app/Http/Controllers/CourseTestimonialMappingController.php
@@ -19,6 +19,9 @@ class CourseTestimonialMappingController extends Controller
     }
 
     public function add(Request $request){
+        if(!in_array('Add Mapping',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $testimonials = Testimonial::all();
         $courses = Course::all();
         $data=[
@@ -29,6 +32,9 @@ class CourseTestimonialMappingController extends Controller
     }
 
     public function edit(Request $request, $id) {
+        if(!in_array('Edit Mapping',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $mappings = CourseTestimonialMapping::find($id);
         $testimonials = Testimonial::all();
         $courses = Course::all();
@@ -99,6 +105,9 @@ class CourseTestimonialMappingController extends Controller
 
     public function destroy($id)
     {
+        if(!in_array('Delete Mapping',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $mapping = CourseTestimonialMapping::find($id);
         $mapping->delete();
         return redirect()->back()->with('msg', 'Mapping Deleted Successfully!');
