+++ app/Http/Controllers/MenuController.php
@@ -15,10 +15,16 @@ class MenuController extends Controller
     }
 
     public function add(Request $request){
+        if(!in_array('Add Menu',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         return view('cms.menu.add');
     }
 
     public function edit(Request $request, $id) {
+        if(!in_array('Edit Menu',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $menus = Menu::find($id);
         $data=[
             'menus' => !empty($menus) ? $menus : []
@@ -94,6 +100,9 @@ class MenuController extends Controller
 
     public function destroy($id)
     {
+        if(!in_array('Delete Menu',\Auth::user()->permissions)){
+            return redirect()->back()->with('error', 'Permission Denied');
+        }
         $menu = Menu::find($id);
         $menu->delete();
         return redirect()->back()->with('msg', 'Menu Deleted Successfully!');
